Move views to edit requirements to app.main.views.create_a_requirement.edit
