+++ app/main/views/buyers.py
@@ -1,14 +1,13 @@
 # coding: utf-8
 from __future__ import unicode_literals
 
-from flask import abort, request, redirect, url_for
+from flask import abort, request, url_for
 from flask_login import current_user
 
 from app import data_api_client
 from .. import main, content_loader
 from ..helpers.buyers_helpers import (
     add_unanswered_counts_to_briefs,
-    brief_can_be_edited,
     count_unanswered_questions,
     get_framework_and_lot,
     is_brief_correct,
@@ -16,10 +15,8 @@ from ..helpers.buyers_helpers import (
     section_has_at_least_one_required_question,
 )
 
-from dmapiclient import HTTPError
 from dmutils.flask import timed_render_template as render_template
 from dmutils.formats import DATETIME_FORMAT
-from dmutils.forms.errors import govuk_errors
 from datetime import datetime
 
 from collections import Counter
@@ -177,131 +174,6 @@ def view_brief_overview(framework_slug, lot_slug, brief_id):
         awarded_brief_response_supplier_name=awarded_brief_response_supplier_name,
         publish_requirements_section_links=publish_requirements_section_links
     ), 200
-
-
-@main.route('/frameworks/<framework_slug>/requirements/<lot_slug>/<brief_id>/<section_slug>', methods=['GET'])
-def view_brief_section_summary(framework_slug, lot_slug, brief_id, section_slug):
-    get_framework_and_lot(
-        framework_slug,
-        lot_slug,
-        data_api_client,
-        allowed_statuses=['live', 'expired'],
-        must_allow_brief=True
-    )
-    brief = data_api_client.get_brief(brief_id)["briefs"]
-
-    if not is_brief_correct(brief, framework_slug, lot_slug, current_user.id) or not brief_can_be_edited(brief):
-        abort(404)
-
-    content = content_loader.get_manifest(brief['frameworkSlug'], 'edit_brief').filter({'lot': brief['lotSlug']})
-    sections = content.summary(brief)
-    section = sections.get_section(section_slug)
-
-    if not section:
-        abort(404)
-
-    # Show preview link if all mandatory questions have been answered
-    unanswered_required, unanswered_optional = count_unanswered_questions(sections)
-    show_dos_preview_link = (unanswered_required == 0)
-
-    return render_template(
-        "buyers/section_summary.html",
-        brief=brief,
-        section=section,
-        show_dos_preview_link=show_dos_preview_link
-    ), 200
-
-
-@main.route(
-    '/frameworks/<framework_slug>/requirements/<lot_slug>/<brief_id>/edit/<section_slug>/<question_id>',
-    methods=['GET'])
-def edit_brief_question(framework_slug, lot_slug, brief_id, section_slug, question_id):
-    get_framework_and_lot(framework_slug, lot_slug, data_api_client, allowed_statuses=['live'], must_allow_brief=True)
-    brief = data_api_client.get_brief(brief_id)["briefs"]
-
-    if not is_brief_correct(brief, framework_slug, lot_slug, current_user.id) or not brief_can_be_edited(brief):
-        abort(404)
-
-    content = content_loader.get_manifest(brief['frameworkSlug'], 'edit_brief').filter(
-        {'lot': brief['lotSlug']}
-    )
-    section = content.get_section(section_slug)
-    if section is None or not section.editable:
-        abort(404)
-
-    question = section.get_question(question_id)
-    if not question:
-        abort(404)
-
-    return render_template(
-        "buyers/edit_brief_question.html",
-        brief=section.unformat_data(brief),
-        section=section,
-        question=question
-    ), 200
-
-
-@main.route(
-    '/frameworks/<framework_slug>/requirements/<lot_slug>/<brief_id>/edit/<section_id>/<question_id>',
-    methods=['POST'])
-def update_brief_submission(framework_slug, lot_slug, brief_id, section_id, question_id):
-    get_framework_and_lot(framework_slug, lot_slug, data_api_client, allowed_statuses=['live'], must_allow_brief=True)
-    brief = data_api_client.get_brief(brief_id)["briefs"]
-
-    if not is_brief_correct(brief, framework_slug, lot_slug, current_user.id) or not brief_can_be_edited(brief):
-        abort(404)
-
-    content = content_loader.get_manifest(brief['frameworkSlug'], 'edit_brief').filter({'lot': brief['lotSlug']})
-    section = content.get_section(section_id)
-    if section is None or not section.editable:
-        abort(404)
-
-    question = section.get_question(question_id)
-    if not question:
-        abort(404)
-
-    update_data = question.get_data(request.form)
-
-    try:
-        data_api_client.update_brief(
-            brief_id,
-            update_data,
-            updated_by=current_user.email_address,
-            page_questions=question.form_fields
-        )
-    except HTTPError as e:
-        update_data = section.unformat_data(update_data)
-        errors = govuk_errors(section.get_error_messages(e.message))
-
-        # we need the brief_id to build breadcrumbs and the update_data to fill in the form.
-        brief.update(update_data)
-
-        return render_template(
-            "buyers/edit_brief_question.html",
-            brief=brief,
-            section=section,
-            question=question,
-            errors=errors
-        ), 400
-
-    if section.has_summary_page:
-        return redirect(
-            url_for(
-                ".view_brief_section_summary",
-                framework_slug=brief['frameworkSlug'],
-                lot_slug=brief['lotSlug'],
-                brief_id=brief['id'],
-                section_slug=section.slug)
-        )
-
-    return redirect(
-        url_for(
-            ".view_brief_overview",
-            framework_slug=brief['frameworkSlug'],
-            lot_slug=brief['lotSlug'],
-            brief_id=brief['id']
-        )
-    )
 
 
 @main.route('/frameworks/<framework_slug>/requirements/<lot_slug>/<brief_id>/responses', methods=['GET'])
