+++ app/main/views/create_a_requirement/__init__.py
@@ -1,3 +1,4 @@
 from . import create
 from . import delete
+from . import edit
 from . import publish
